+++ test/cli/test_main.py
@@ -41,7 +41,7 @@ def test_help(
     assert "usage: whl2conda2" in out
     assert "--markdown-help" not in out
 
-    subcmds = ["build", "convert", "config", "diff", "install"]
+    subcmds = ["build", "convert", "config", "diff", "install", "test"]
     for subcmd in subcmds:
         assert re.search(rf"^\s+{subcmd}\s+\w+", out, flags=re.MULTILINE)